Make damage values a public variable. Allow bots to die from bullet damage.

+++ Assets/Scripts/PlayerController.cs
@@ -179,7 +179,7 @@ namespace Assets.Scripts
             NetworkServer.Spawn(mine.gameObject);
         }
 
-        public bool HitByBullet(int damage, short owner)
+        public bool HitByBullet(float damage, short owner)
         {
             if (isLocalPlayer)
                 CmdTakeDamage(damage, owner);
